+++ src/QuickMarcEditor/QuickMarcEditor.js
@@ -117,7 +117,7 @@ const QuickMarcEditor = ({
   };
 
   if ((marcType === MARC_TYPES.AUTHORITY) && records.length) {
-    recordInfoProps.correspondingMarcTag = getCorrespondingMarcTag(records);
+    recordInfoProps.correspondingMarcTag = getCorrespondingMarcTag(initialValues.records);
   }
 
   const getPaneTitle = () => {
@@ -134,12 +134,18 @@ const QuickMarcEditor = ({
       }
 
       if ((marcType === MARC_TYPES.AUTHORITY) && records.length) {
-        const headingTagContent = records.find((recordRow) => {
+        const currentHeading = records.find((recordRow) => {
           return recordRow.tag === recordInfoProps.correspondingMarcTag;
-        }).content;
+        });
+
+        const initialHeading = initialValues.records.find((recordRow) => {
+          return recordRow.tag === recordInfoProps.correspondingMarcTag;
+        });
+
+        const headingContent = currentHeading?.content || initialHeading?.content;
 
         formattedMessageValues = {
-          title: getContentSubfieldValue(headingTagContent).$a,
+          title: getContentSubfieldValue(headingContent).$a,
         };
       } else {
         formattedMessageValues = instance;